+++ mangareaderapi/admin/routes/author.py
@@ -2,10 +2,11 @@ from flask import request, jsonify, Blueprint
 from mangareaderapi import app, db
 from mangareaderapi.models import Author
 from mangareaderapi.schema import author_schema, authors_schema
+from mangareaderapi.admin.routes.utils import check_existing_by_name
 
 author = Blueprint("author", __name__)
 
-# Author
+
 @author.route("/author", methods=["GET"])
 def get_authors():
     all_authors = Author.query.all()
@@ -25,32 +26,41 @@ def get_author(id):
 def add_author():
     name = request.json["name"]
 
-    new_author = Author(name)
+    if check_existing_by_name(name, Author):
+        return jsonify(message=f"The author is already existing")
+    else:
+        new_author = Author(name)
 
-    db.session.add(new_author)
-    db.session.commit()
+        db.session.add(new_author)
+        db.session.commit()
 
-    return author_schema.jsonify(new_author)
+        return author_schema.jsonify(new_author)
 
 
 @author.route("/author/<id>", methods=["PUT"])
 def update_author(id):
     author = Author.query.get(id)
 
-    name = request.json["name"]
+    if author:
+        name = request.json["name"]
 
-    author.name = name
+        author.name = name
 
-    db.session.commit()
+        db.session.commit()
 
-    return author_schema.jsonify(author)
+        return author_schema.jsonify(author)
+    else:
+        return jsonify(message=f"The author does not exist")
 
 
 @author.route("/author/<id>", methods=["DELETE"])
 def delete_author(id):
     author = Author.query.get(id)
 
-    db.session.delete(author)
-    db.session.jsonify(author)
+    if author:
+        db.session.delete(author)
+        db.session.commit()
 
-    return author_schema.jsonify(author)
+        return author_schema.jsonify(author)
+    else:
+        return jsonify(message="The author does not exist")
