+++ mangareaderapi/admin/routes/genre.py
@@ -2,10 +2,11 @@ from flask import request, jsonify, Blueprint
 from mangareaderapi import app, db
 from mangareaderapi.models import Genre
 from mangareaderapi.schema import genre_schema, genres_schema
+from mangareaderapi.admin.routes.utils import check_existing_by_name
 
-genre = Blueprint('genre', __name__)
+genre = Blueprint("genre", __name__)
 
-# Genre Routes
+
 @genre.route("/genre", methods=["GET"])
 def get_genres():
     all_genres = Genre.query.all()
@@ -23,33 +24,42 @@ def get_genre(id):
 def add_genre():
     name = request.json["name"]
 
-    new_genre = Genre(name)
+    if check_existing_by_name(name, Genre):
+        return jsonify(message="The genre is already existing")
+    else:
+        new_genre = Genre(name)
 
-    db.session.add(new_genre)
-    db.session.commit()
+        db.session.add(new_genre)
+        db.session.commit()
 
-    return genre_schema.jsonify(new_genre)
+        return genre_schema.jsonify(new_genre)
 
 
 @genre.route("/genre/<id>", methods=["PUT"])
 def update_genre(id):
     genre = Genre.query.get(id)
 
-    name = request.json["name"]
+    if genre:
+        name = request.json["name"]
 
-    genre.name = name
+        genre.name = name
 
-    db.session.commit()
+        db.session.commit()
 
-    return genre_schema.jsonify(genre)
+        return genre_schema.jsonify(genre)
+    else:
+        return jsonify(message="The genre does not exist")
 
 
 @genre.route("/genre/<id>", methods=["DELETE"])
 def delete_genre(id):
     genre = Genre.query.get(id)
 
-    db.session.delete(genre)
-    db.session.commit()
+    if genre:
+        db.session.delete(genre)
+        db.session.commit()
 
-    return genre_schema.jsonify(genre)
+        return genre_schema.jsonify(genre)
+    else:
+        return jsonify(message="The genre does not exist")
